creating and deleting comment

+++ controllers/postController.js
@@ -1,13 +1,27 @@
 const Post = require("../models/post");
 const User = require("../models/user");
+const Comment = require("../models/comment");
 const fs = require("fs");
 const path = require("path");
 
-module.exports.getPosts = async (req, res) => {
+module.exports.getUserPosts = async (req, res) => {
   try {
-    let posts = await Post.find({ user: req.user.id }).sort("-createdAt");
+    let posts = await Post.find({ user: req.user.id })
+      .sort("-createdAt")
+      .populate("user");
 
-    return res.render("profile_post", { layout: false, posts: posts });
+    return res.render("post", { layout: false, posts: posts });
+  } catch (error) {
+    console.log(error);
+    return "";
+  }
+};
+
+module.exports.getHomePosts = async (req, res) => {
+  try {
+    let posts = await Post.find().sort("-createdAt").populate("user");
+
+    return res.render("post", { layout: false, posts: posts });
   } catch (error) {
     console.log(error);
     return "";
@@ -41,8 +55,8 @@ module.exports.createPost = async (req, res) => {
               message: "Post created",
             });
           }
-        } catch (err) {
-          console.log(err);
+        } catch (error) {
+          console.log(error);
           return res.redirect("back");
         }
       })();
@@ -54,27 +68,33 @@ module.exports.deletePost = async (req, res) => {
   // console.log(req.params.id);
   try {
     let post = await Post.findById(req.params.id);
-    if (post){ 
+    if (post && post.user == req.user.id) {
       await fs.unlinkSync(path.join(__dirname, "..", post.content));
       await post.remove();
+      await Comment.deleteMany({post: req.params.id});
+      if (req.xhr) {
+        return res.status(200).json({
+          data: {
+            post_id: req.params.id,
+          },
+          message: "Post deleted",
+        });
+      }
     }
-    if (req.xhr) {
-      return res.status(200).json({
-        data: {
-          post_id: req.params.id,
-        },
-        message: "Post deleted",
-      });
-    }
+    return res.status(400).json({
+      message: "Post could not be deleted",
+    });
   } catch (error) {
-    // console.log(error);
-    return error;
+    console.log(error);
+    return res.status(500).json({
+      message: "Post could not be deleted",
+    });
   }
 };
 
 module.exports.reactions = (req, res) => {
   //   console.log(req);
-  return res.render("reactions", {
+  return res.render("reaction", {
     layout: false,
     title: "Reactions",
     // postid: req.body.postid,
@@ -102,12 +122,5 @@ module.exports.reactions = (req, res) => {
         { display_pic: "/images/home-page.jpg", user_name: "Sita" },
       ],
     },
-  });
-};
-
-module.exports.comments = (req, res) => {
-  return res.render("comments", {
-    layout: false,
-    title: "Comments",
   });
 };
